requirements.txt
fastapi
firebase-admin
python-dotenv
uvicorn
python-jose
